pyproject.toml
[project]
name = "lehrer-lyrics"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.135.3",
    "httpx>=0.28.1",
    "markdown>=3.10.2",
    "pydantic>=2.12.5",
    "python-fasthtml>=0.13.2",
]

[build-system]
requires = ["uv_build>=0.11.3,<0.12"]
build-backend = "uv_build"

[project.scripts]
lehrer-scrape = "lehrer_lyrics.scraper.cli:app"

[tool.fastapi]
entrypoint = "lehrer_lyrics.service.main:app"

[dependency-groups]
dev = [
    "beautifulsoup4>=4.14.3",
    "notebook>=7.5.5",
    "ollama>=0.6.1",
    "pypdf>=6.9.2",
    "pytest>=9.0.2",
    "pytest-cov>=7.1.0",
    "pytest-xdist>=3.8.0",
    "rich>=14.3.3",
    "ruff>=0.15.8",
    "ty>=0.0.27",
    "typer>=0.24.1",
]
